// src/app/shared/model/new-cust/shareholder/shareholder-listing-obj.model.ts
export class ShareholderListingObj{
    CustId: number;
    ShareholderId: number;
    CustCompanyMgmntShrholderId: number;
    CustNo: string;
    ShareholderName: string;
    ShareholderType: string;
    ShareholderTypeDesc: string;
    MrPublicTypeCode: string;
    MrPublicTypeDesc: string;
    SharePrcnt: number;
    IsSigner: boolean;
    IsOwner: boolean;
    IsActive: boolean;
    NegCustTypeCode: string;
    NegCustTypeCodeDesc: string;
    IsNeedInputSpouseShareholderData: boolean;
    constructor(){
        
    }
}